+++ NEWS.md
@@ -1,4 +1,8 @@
+# ps (development version)
+
 # ps 1.7.5
+
+No user visible changes.
 
 # ps 1.7.4