+++ NEWS.md
@@ -1,5 +1,7 @@
 # ps (development version)
 
+* New function `ps_disk_io_counters()` to query disk I/O counters implemented on Windows and Linux. (#144, @michaelwalshe).
+
 # ps 1.7.5
 
 No user visible changes.