+++ NEWS.md
@@ -1,5 +1,7 @@
 # ps (development version)
 
+* `ps_name()` now does not fail in the rare case when `ps_cmdline()` returns an empty vector (#150).
+
 * `ps_system_cpu_times()` now returns CPU times divided by the HZ as reported by CLK_TCK, in-line with other OS's and the per-process version. (#144, @michaelwalshe).
 
 # ps 1.7.5