+++ NEWS.md
@@ -1,4 +1,4 @@
-# ps (development version)
+# ps 1.7.6
 
 * `ps_name()` now does not fail in the rare case when `ps_cmdline()` returns an empty vector (#150).